GITHUB-APD-162: Rework unit front tests
