--- src/Akeneo/Platform/Bundle/CommunicationChannelBundle/Resources/workspaces/communication-channel/tests/front/unit/__mocks__/oro/mediator.ts
@@ -1,3 +1,0 @@
-export = {
-  trigger: jest.fn((key: string) => key),
-};
