+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/legacy-bridge/src/index.ts
@@ -1,4 +1,3 @@
 export * from './components';
 export * from './hooks';
 export * from './provider';
-export * from './bridge/react';
