+++ tests/front/unit/jest/fetchMock.ts
@@ -1,0 +1,5 @@
+import {GlobalWithFetchMock} from 'jest-fetch-mock';
+
+const customGlobal: GlobalWithFetchMock = global as GlobalWithFetchMock;
+customGlobal.fetch = require('jest-fetch-mock');
+customGlobal.fetchMock = customGlobal.fetch;
